# docs/Worth_It_OpenAPI_v1.0.yaml
openapi: 3.1.0
info:
  title: Worth It API
  version: 1.0.0
  description: Internal REST API contract for the Worth It mobile app MVP. All monetary amounts use integer minor
    units plus an ISO 4217 currency code.
servers:
- url: https://api.worthit.app/v1
  description: Production
- url: https://staging-api.worthit.app/v1
  description: Staging
security:
- bearerAuth: []
...
paths:
  /me:
    get:
      summary: Get current account
      tags:
      - Account
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /me/settings:
    patch:
      summary: Update account settings
      tags:
      - Account
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                preferred_currency:
                  type: string
                notifications_enabled:
                  type: boolean
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /account-deletion:
    post:
      summary: Request account deletion
      tags:
      - Account
      responses:
        '202':
          description: Deletion accepted
        '401':
          description: Unauthorized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - confirmation
              properties:
                confirmation:
                  type: string
                  const: DELETE
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
  /ad-rewards/challenges:
    post:
      summary: Create rewarded-ad verification challenge
      tags:
      - Ads
      responses:
        '201':
          description: Challenge created
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - provider
              properties:
                provider:
                  type: string
                placement:
                  type: string
                  const: scan_unlock
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /ad-rewards/claims:
    post:
      summary: Claim one scan after verified ad completion
      tags:
      - Ads
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - challenge_id
              - provider_transaction_id
              properties:
                challenge_id:
                  type: string
                  format: uuid
                provider_transaction_id:
                  type: string
                verification_payload:
                  type: object
                  additionalProperties: true
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /subscription:
    get:
      summary: Get subscription entitlement
      tags:
      - Subscriptions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
                  is_active:
                    type: boolean
                  expires_at:
                    type: string
                    format: date-time
                  store:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /subscription/restore:
    post:
      summary: Reconcile restored store purchases
      tags:
      - Subscriptions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  reconciled:
                    type: boolean
                  plan:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                app_user_id:
                  type: string
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
...
  /webhooks/ad-provider:
    post:
      summary: Receive ad provider server-side verification
      tags:
      - Webhooks
      responses:
        '204':
          description: Accepted
        '401':
          description: Invalid signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      security:
      - webhookSignature: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  schemas:
...
    Error:
      type: object
      required:
      - error
      - request_id
      properties:
        error:
          type: object
          required:
          - code
          - message
          - retryable
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    User:
      type: object
      required:
      - id
      - preferred_currency
      - plan
      - scan_balance
      properties:
        id:
          type: string
          format: uuid
        preferred_currency:
          type: string
        plan:
          type: string
          enum:
          - free
          - premium
        scan_balance:
          type: object
          properties:
            introductory_remaining:
              type: integer
            rewarded_remaining:
              type: integer
            unlimited:
              type: boolean
        created_at:
          type: string
          format: date-time
...